requirements.txt
asgiref==3.2.3
boto3==1.12.17
botocore==1.15.17
Django==3.0.4
docutils==0.15.2
image==1.5.28
jmespath==0.9.5
Pillow==7.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
